Make MoleculeSubForm smaller, by moving reusable components to own files
