+++ app/scenarios/HiddenFileInput.tsx
@@ -1,0 +1,15 @@
+import { useEffect, useRef } from "react";
+
+export function HiddenFileInput({ name, file }: { name: string; file: File }) {
+  const ref = useRef<HTMLInputElement>(null);
+
+  useEffect(() => {
+    if (ref.current) {
+      const dataTransfer = new DataTransfer();
+      dataTransfer.items.add(file);
+      ref.current.files = dataTransfer.files;
+    }
+  }, [file]);
+
+  return <input ref={ref} type="file" name={name} className="hidden" />;
+}
